fix(process): fix unexpected panic when spawn time was in future
A panic occurred after a process's calculated spawn time was greater than the current timestamp. This caused a panic later on when updating the process's running span, as its new end was now less than the span's begin.

The applied fix is to check the spawn time of a process when it is being built, and replace it with the current time if the calculated time was greater than now.