Render process span (#13)
* feat(sysconf): add module to get sysconf values

* refactor(parsers): rename test_utils mod fakes

* refactor(parsers): rename Data trait Parse

* feat(parsers): add Parse structs to calculate proc uptime

* feat(process): calculate process spawn time when fetching metadata

* fix(test): add missing spawn time param to ProcessMetadata

* doc(PidStat): explain unit of PidStat::starttime()

* fix(ctrl/proc): keep tracking initially selected process

* feat(ui): display process span on metadata bar

* refactor(terminal): inline render_on_frame

* fix(process): fix unexpected panic when spawn time was in future

A panic occurred after a process's calculated spawn time was greater than the current timestamp. This caused a panic later on when updating the process's running span, as its new end was now less than the span's begin.

The applied fix is to check the spawn time of a process when it is being built, and replace it with the current time if the calculated time was greater than now.

* fix(ui): stop displaying process status

The process status was displayed in the selected process' metadata bar. It has been removed as it was redundant with the process' running span which is now also displayed in this area.