tempfix(rates): prevent crash and add logs when error occurs in rate
The error occurred when calculating disk IO rates. For a given PID, a new value was less than a past value. As values used to calculate rates are expected to always be increasing, this provoked a subtraction overflow error.

The root cause for this issue is not clear yet (reassigned pid, io counter which may decrease in procfs, ...). To be able to debug this issue in the future, an error log is now emitted when this problem occurs.

To prevent a crash in such cases, we use saturating_sub() instead of the subtraction operator.