feat(ctrl): allow different behavior for same input depending on state
Depending on the state of the application, the same input can be used for different behaviors.

In the nominal state of the application, the up/down key will select the next/previous process in the process list.

When displaying the process ordering criteria pop-up, the up/down key will select the criteria to apply.

This input interpretation has been moved to the ctrl module, as it became too complex to be left in the spv module.