fix: chart now display line across whole screen even on inaccurate span
When the rendering span's bounds did not align with dated metrics, the edges of the chart did not display any metrics. That is because the first metric in the span was slightly after the left edge, and the last metric slightly before the right edge.

This has been fixed by including metrics adjacent to the span in ProcessData::view()